Move `stubPrecompileEnvironment` to `legacy_test.go`
